Improve consent management authorization handling
Introduce a new consent manager impl to be consumed as an osgi service free of authorization.
Fixes https://github.com/wso2/product-is/issues/4311
